+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
+## [1.2.0] - 28 Jun, 2022
+- Introduces the new "Web Personalization" channel with support for key-value templates.
+- Adds new method renderNotifcationViewed to capture Notification Viewed events.
+- Adds new method renderNotificationClicked to capture Notification Clicked events.
+- Adds analytics support for upcoming CleverTap Direct Call Web SDK.
 
 ## [1.1.3] - 14 Jun, 2022
 - Fixed a bug where multiple CleverTap IDs were generated due to a variable name mismatch.